reduce unchecked integer casts with `as` (#431)
In general, determining if an `as` cast is safe requires close
examination.  To reduce this cognitive load, we should adopt conventions
around use of `as` so that we don't have to think about the large
majority of conversions.

The goals of this commit are:

* To always perform runtime checking of any conversion that could change
the resulting value (e.g. usize -> u32).

* To make it easier to inspect the code and manually verify that the
above rule is followed.

To maximize compliance with the conventions, we'll enable these clippy
lints, and not have any additional requirements:

#![warn(clippy::cast_lossless)]
#![warn(clippy::cast_possible_truncation)]
#![warn(clippy::cast_possible_wrap)]
#![warn(clippy::cast_sign_loss)]

In particular, this allows `usize as u64` but not `u64 as usize`, for
which we will use the new `usize::from64(...)` trait.

This commit makes a few changes:

* Change some struct member types to reduce the amount of
casting/conversion necessary.

* Use `from()` where possible - for conversions that can never lose
state (e.g. u32 -> u64).

* Use `try_from().unwrap()` whenever a conversion could fail, given our
platform (LP64) (e.g. usize -> u32)

* For conversions that can't fail on our platform, but which the clippy
lints flag (specifically just u64 -> usize), we will use the new
`usize::from64(...)` function.

* For other conversions (i.e. all those not flagged by the enabled
clippy lints), use `as`, e.g. usize -> u64, u32 -> u64, u32->usize.

+++ cmd/zfs_object_agent/zettacache/src/space_map.rs
@@ -110,7 +110,7 @@ impl SpaceMap {
             location: DiskLocation {
                 offset: self.coverage.offset,
             },
-            size: self.coverage.size as usize,
+            size: self.coverage.size,
         }
     }
 }
